% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_country.R
\name{compute_country_ne}
\alias{compute_country_ne}
\title{Title}
\usage{
compute_country_ne(data, scales, keep_country = NULL)
}
\arguments{
\item{keep_country}{}
}
\description{
Title
}
\examples{
library(dplyr)
#northcarolina_flat |> rename(fips = FIPS) |> compute_country_northcarolina() |> head()
#northcarolina_flat |> rename(fips = FIPS) |> compute_country_northcarolina(keep_country = "Ashe")
}
