% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_county.R
\name{compute_county_northcarolina}
\alias{compute_county_northcarolina}
\title{Title}
\usage{
compute_county_northcarolina(data, scales, keep_county = NULL)
}
\arguments{
\item{keep_county}{}
}
\description{
Title
}
\examples{
library(dplyr)
#northcarolina_flat |> rename(fips = FIPS) |> compute_county_northcarolina() |> head()
#northcarolina_flat |> rename(fips = FIPS) |> compute_county_northcarolina(keep_county = "Ashe")
}
