% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_country_labels.R
\name{compute_panel_country_centers}
\alias{compute_panel_country_centers}
\title{Title}
\usage{
compute_panel_country_centers(data, scales, keep_country = NULL)
}
\arguments{
\item{keep_country}{}
}
\description{
Title
}
