% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_county.R
\name{geom_county}
\alias{geom_county}
\title{Title}
\usage{
geom_county(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  crs = "NAD27",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
